+++ backend/producto.py
@@ -1,10 +1,9 @@
-from backend.hoja_cliente import obtenerHojaDeClientes
+from backend.hoja_producto import obtenerHojaDeProductos
 
-hoja = obtenerHojaDeClientes()
+hoja = obtenerHojaDeProductos()
 
-def listarClientes():
+def listarProductos():
     filas = []
-
     refFilas = hoja.iter_rows(min_row=2, max_row=hoja.max_row, min_col=1, max_col=4)
 
     for refFila in refFilas:
@@ -15,12 +14,12 @@ def listarClientes():
 
     return filas
 
-def consultarCliente(documento):
+def consultarProducto(id_producto):
     refFilas = hoja.iter_rows(min_row=2, max_row=hoja.max_row, min_col=1, max_col=4)
     refFilasEnum = enumerate(refFilas)
 
     for idx, refFila in refFilasEnum:
-        if refFila[0].value == documento:
+        if refFila[0].value == id_producto:
             valores = [idx]
             for celda in refFila:
                 valores.append(celda.value)
