--- main.py
@@ -1,5 +1,0 @@
-from backend.cliente import listarClientes
-
-clientes = listarClientes()
-
-print(clientes)
