--- main2.py
@@ -1,5 +1,0 @@
-from backend.cliente import consultarCliente
-
-cliente = consultarCliente("1234")
-
-print(cliente)
